+++ backend/entityservice/default_logging.yaml
@@ -47,28 +47,27 @@ handlers:
 
 loggers:
   entityservice:
-    level: DEBUG
-    propagate: yes
-
-  entityservice.tasks:
-    level: DEBUG
-    propagate: no
+    level: INFO
 
   entityservice.database.util:
     level: WARNING
-    propagate: no
+
+  entityservice.cache:
+    level: WARNING
+
+  entityservice.utils:
+    level: INFO
 
   celery:
     level: INFO
-    propogate: yes
 
   jaeger_tracing:
     level: WARNING
-    propogate: no
+    propagate: no
 
   werkzeug:
-    level: WARN
-    propogate: no
+    level: WARNING
+    propagate: no
 
 root:
   level: INFO
