+++ backend/entityservice/verbose_logging.yaml
@@ -58,6 +58,10 @@ loggers:
     level: INFO
     propagate: yes
 
+  entityservice.cache:
+    level: DEBUG
+    propagate: yes
+
   celery:
     level: WARNING
     propogate: yes
