+++ backend/entityservice/views/run/status.py
@@ -2,7 +2,8 @@ from flask import request, g
 from structlog import get_logger
 import opentracing
 
-from entityservice import database as db, cache as cache
+from entityservice.cache import progress as progress_cache
+from entityservice import database as db
 from entityservice.views.auth_checks import abort_if_run_doesnt_exist, get_authorization_token_type_or_abort
 from entityservice.views.serialization import completed, running, error
 from entityservice.models.run import RUN_TYPES
@@ -11,7 +12,7 @@ logger = get_logger()
 
 
 def get(project_id, run_id):
-    log = logger.bind(pid=project_id,rid=run_id)
+    log = logger.bind(pid=project_id, rid=run_id)
     parent_span = g.flask_tracer.get_span()
     log.debug("request run status")
     with opentracing.tracer.start_span('check-auth', child_of=parent_span) as span:
@@ -48,7 +49,7 @@ def get(project_id, run_id):
             max_val = db.get_project_column(conn, project_id, 'parties')
     elif stage == 2:
         # Computing similarity
-        abs_val = cache.get_progress(run_id)
+        abs_val = progress_cache.get_progress(run_id)
         if abs_val is not None:
             with db.DBConn() as conn:
                 max_val = db.get_total_comparisons_for_project(conn, project_id)
